+++ src/drtsans/mono/gpsans/api.py
@@ -1147,6 +1147,9 @@ def reduce_single_configuration(loaded_ws, reduction_input, prefix="", skip_nan=
             "background_width": reduction_config["autoWedgeBackgroundWidth"],
             "signal_to_noise_min": reduction_config["autoWedgeSignalToNoiseMin"],
             "peak_search_window_size_factor": reduction_config["autoWedgePeakSearchWindowSizeFactor"],
+            "auto_wedge_phi_min": reduction_config["autoWedgePhiMin"],
+            "auto_wedge_phi_max": reduction_config["autoWedgePhiMax"],
+            "auto_symmetric_wedges": reduction_config["autoSymmetricWedges"],
         }
         # auto-aniso returns all of the wedges
         symmetric_wedges = False