+++ docs/release_notes.rst
@@ -26,6 +26,7 @@ Release Notes
 - MR #1169: I(Qx, Qy) is saved in NXCANSAS format for biosans, gpsans, and eqsans workflows
 - MR #1168: Document scaling the detector panels in the user's guide
 - MR #1167: Allow separate configurations for inelastic incoherence correction per frame in frame skipping mode
+- MR #1166: Added option to _fitSpectrum to auto-find one wedge and mirror it
 - MR #1162: When reducing `gpsans` data with `direct_beam` scaling, the `direct_beam_scaling` parameter is now logged during
   the reduction process and stored in the output Nexus file at `reduction_information/special_parameters/direct_beam_scaling/value`.
 - MR #1161: Add a parameters  removeAlgorithmHistory to write less data and speed up I/O during reduction